+++ CHANGELOG.md
@@ -5,3 +5,5 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED
 -----------------
 - Fixed Fireflies not rendering their base texture.
 - Fixed Display Lanterns not properly rendering their contents when updated.
+- Added the Reach Boost effect to Beacons, thanks to a suggestion by wet.noodle/Ashlyn.
+  - Added the `Reach Boost Beacons` config to toggle this.